Moved app import util into the hook
